desktop/requirements.txt
# PatentFlow Desktop Requirements
# Core dependencies for PyQt6 desktop application

# UI Framework
PyQt6==6.6.1
PyQt6-tools==6.6.1

# Word Integration
pywin32==306

# NLP Processing
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz

# Graph Analysis
networkx==3.2.1
matplotlib==3.8.2

# Text Processing
regex==2023.10.3
python-Levenshtein==0.23.0

# Async & Threading
asyncio==3.4.3

# Configuration & Logging
pyyaml==6.0.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-qt==4.2.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API Client (for backend communication)
requests==2.31.0
aiohttp==3.9.1

# Data handling
pandas==2.1.3
numpy==1.25.2